requirements.txt
pandas~=2.2.3
requests~=2.32.3
fastapi~=0.115.11
strawberry-graphql~=0.262.5
SQLAlchemy~=2.0.39